* feat(Sign.vue): refactor Sign component to use SignLoader component for displaying migration steps * feat(SignLoader.vue): create SignLoader component to display migration steps * fix(utils.ts): add type definition for Steps
